Refactor: Enhance Detail Screen and UI/UX
This commit significantly overhauls the Detail Screen, refactors the Main Screen, and improves the overall user experience and data presentation.

- **Data Layer:**
    - `MainMapper.kt`:
        - Adjusted `height` and `weight` to be `Float` values by dividing the original integer values by 10.
        - Improved stat name mapping for better display (e.g., "special-attack" to "Sp. Attack").
        - Implemented `formatGeneration` to properly format the generation string (e.g., "generation-i" to "Generation I").

- **Domain Layer:**
    - `PokemonModel.kt`: Updated `height` and `weight` fields from `Int` to `Float` to reflect the new data format.

- **App Layer:**
    - **Detail Screen Overhaul:**
        - Introduced a tab-based layout with "Information", "Base Stats", and "Evolutions" sections.
        - Created `DetailTab` enum and updated `DetailState` and `DetailViewModel` to manage tab selection.
        - `PokemonInfoSection.kt`: New composable to display Pokémon's "About" info (weight, height, etc.), gender distribution, and abilities.
        - `PokemonStatsSection.kt`: New composable to visualize base stats with progress bars, color-coded based on value, and a total calculation.
        - `PokemonEvolutionSection.kt`: New composable to display the Pokémon's evolution line with sprites, names, and IDs, connected by a Lottie animation.
        - `HeroSection.kt`: Refactored the top part of the Detail Screen into its own composable.
        - Added extensive string resources for all new UI text.

    - **Main Screen UI/UX Improvements:**
        - Refactored `PokemonList` and `PokemonItem` into their own file (`app/.../main/components/PokemonList.kt`).
        - Replaced the `TopAppBar` with a cleaner title and description section.
        - Enhanced the search field with a placeholder and a leading search icon.
        - Updated string resources for titles, descriptions, and button labels to improve clarity.

    - **Navigation & Component Refactoring:**
        - Removed `HomeScreen` and its `ViewModel`, making `MainScreen` the starting destination.
        - Removed `GuideScreen` and its `ViewModel`.
        - Created `Lotties.kt` to centralize Lottie animation composables (`LottieArrowRight`, `LottiePokeball`).

    - **Branding and Assets:**
        - Added new app logo assets (`ic_logo*`) and updated `AndroidManifest.xml` to use them.

    - **Code Cleanup:**
        - Removed unused imports and cleaned up test files.
        - Simplified `MainActivity.kt` by removing unused composables.

+++ domain/src/main/java/com/hugopolog/domain/entities/pokemon/PokemonModel.kt
@@ -3,8 +3,8 @@ package com.hugopolog.domain.entities.pokemon
 data class PokemonModel(
     val id: Int,
     val name: String,
-    val height: Int,
-    val weight: Int,
+    val height: Float,
+    val weight: Float,
     val gender: GenderDistribution,
     val baseExperience: Int,
     val abilities: List<String>,
